requirement.txt
asgiref==3.8.1
bootstrap4==0.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.3.0
django-configurations==2.5.1
django-cors-headers==3.10.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.0.0
